add `evaluate_with_ops`
the original use-case for this library is to share tree transformations with clients over a websocket, by applying operations to a local tree and then sending those same operations to the client to apply locally.

Adding the operations to the output as an additional function makes sense in this case, so that the caller doesn't have to case over the return to build up a new tuple.